Migrate data sinks in brokers-jvm module (#1614)
