--- streampipes-extensions/streampipes-sinks-brokers-jvm/src/main/java/org/apache/streampipes/sinks/brokers/jvm/rabbitmq/RabbitMqConsumer.java
@@ -1,69 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-package org.apache.streampipes.sinks.brokers.jvm.rabbitmq;
-
-import org.apache.streampipes.commons.exceptions.SpRuntimeException;
-import org.apache.streampipes.dataformat.json.JsonDataFormatDefinition;
-import org.apache.streampipes.model.runtime.Event;
-import org.apache.streampipes.pe.shared.PlaceholderExtractor;
-import org.apache.streampipes.wrapper.context.EventSinkRuntimeContext;
-import org.apache.streampipes.wrapper.runtime.EventSink;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-public class RabbitMqConsumer implements EventSink<RabbitMqParameters> {
-
-  private RabbitMqPublisher publisher;
-  private JsonDataFormatDefinition dataFormatDefinition;
-  private String topic;
-
-  private static final Logger LOG = LoggerFactory.getLogger(RabbitMqConsumer.class);
-
-  public RabbitMqConsumer() {
-    this.dataFormatDefinition = new JsonDataFormatDefinition();
-  }
-
-  @Override
-  public void onInvocation(RabbitMqParameters parameters, EventSinkRuntimeContext runtimeContext)
-      throws SpRuntimeException {
-    this.publisher = new RabbitMqPublisher(parameters);
-    this.topic = parameters.getRabbitMqTopic();
-
-    if (!this.publisher.isConnected()) {
-      throw new SpRuntimeException("Could not establish conntection to RabbitMQ broker. Host: "
-          + parameters.getRabbitMqHost() + " Port: " + parameters.getRabbitMqPort());
-    }
-  }
-
-  @Override
-  public void onEvent(Event inputEvent) {
-    try {
-      publisher.fire(dataFormatDefinition.fromMap(inputEvent.getRaw()),
-          PlaceholderExtractor.replacePlaceholders(inputEvent, topic));
-    } catch (SpRuntimeException e) {
-      LOG.error("Could not serialiaze event");
-    }
-  }
-
-  @Override
-  public void onDetach() throws SpRuntimeException {
-    publisher.cleanup();
-  }
-}
